+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 2021-02-08 0.0.6
+
+* added /api/dashboards to gauges
+* added gauges
+  * `foreman_exporter_dashboard_total_hosts`
+  * `foreman_exporter_dashboard_bad_hosts`
+  * `foreman_exporter_dashboard_bad_hosts_enabled`
+  * `foreman_exporter_dashboard_active_hosts`
+  * `foreman_exporter_dashboard_active_hosts_ok`
+  * `foreman_exporter_dashboard_active_hosts_ok_enabled`
+  * `foreman_exporter_dashboard_ok_hosts`
+  * `foreman_exporter_dashboard_ok_hosts_enabled`
+  * `foreman_exporter_dashboard_disabled_hosts`
+  * `foreman_exporter_dashboard_pending_hosts`
+  * `foreman_exporter_dashboard_pending_hosts_enabled`
+  * `foreman_exporter_dashboard_out_of_sync_hosts`
+  * `foreman_exporter_dashboard_out_of_sync_hosts_enabled`
+  * `foreman_exporter_dashboard_good_hosts`
+  * `foreman_exporter_dashboard_good_hosts_enabled`
+  * `foreman_exporter_dashboard_percentage`
+  * `foreman_exporter_dashboard_reports_missing`
+  * `foreman_exporter_dashboard_request_time_seconds`
+* rename `environment` label from hosts to `puppet_environment`)
+* changed `sleep` main loop into threading
+
 ## 2021-02-07 0.0.5
 
 * improved raise exception